refactor: move OpAmp saturator into dsp
